**Changes** *Moved all files out of sub files and into src file. *Added to HeartContentArray.ts.
